修复: GPU蒙皮compute shader输出全零 + 角色穿透地面 + 索引升级uint32
- gpu_skinning.cpp: GL430生成的uniform uint改为uniform int，修复glUniform1i对uint uniform无效导致compute shader所有线程提前退出

- lua_binding_ecs_phys3d.cpp: Jolt MoveCharacter后补查TerrainHeightmapComponent地形贴地

- mesh_render_system.cpp: GPU蒙皮readback路径增加判空保护，无效时回退VS蒙皮

- 全链路mesh索引从unsigned short升级到uint32_t(三后端draw executor+softbody+fracture+grass)

- glfw_app.cpp: 初始化时设置箭头光标避免Windows忙碌光标

+++ engine/ecs/components_3d.h
@@ -65,7 +65,7 @@ struct MeshRendererComponent {
     MaterialDataSource material_data_source = MaterialDataSource::ComponentFallback;
     unsigned int mesh_handle_override = 0;  ///< 非零时表示 LODSystem 正在管理此实体的 mesh_path
     std::vector<float> temp_vertices;
-    std::vector<unsigned short> temp_indices;
+    std::vector<uint32_t> temp_indices;
     std::vector<float> temp_uvs;
     std::vector<float> temp_normals;
     std::vector<float> temp_tangents;
